Move all components back into components.
