+++ src/components/artist/ArtistModal.tsx
