+++ src/components/artist/index.ts
